Completed full join menu functionality
All of the buttons and popups now work!

In addition, some other minor things, including the example servers, the text that appears in some popups, minor code organization, etc.

I really hope nothing is named jankily in the script for the join menu without me realizing it, I don't think so, but a lot of it is copied over from the worlds menu script.