Renamed coords formats
xyz -> zyx
hxx -> hzz

The main issue with xyz format is that it's very difficult to expand upon given that z is the last letter in the english alphabet. (For 4 dimensions, w is awkwards used after z.) By using z first and counting down (z y x w v u ...) this problem is avoided for up to 26 variables, and if you need more than that then perhaps you should be using a different system entirely anyways.