prevented crashes & tooltip experimentation
"prevented crashes" refers to the crash that previously happened if you started the program with the worlds or remembered servers menus as the primary menu visible. Given that that won't happen in normal function, the change probably won't affect players too much, but it's nicer for development testing with the menus.