mouse now returns to previous pos' after capture
When playing in a world, you can right click and hold to rotate the temporary camera view around. Previously, releasing right click when doing this would plop your mouse cursor to the center of the window regardless of where it was when you first started holding right click.

(Also minorly edited credits text.)